add entry search
